+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Fixes
 
+- storage: panic on unexpected state in MemoryStorage ([#75](https://github.com/lamports-dev/alpamayo/pull/75))
+
 ### Features
 
 ### Breaking